lint changes

+++ include/IShellyWrapper.h
@@ -1,15 +1,16 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 #pragma once
+#include <stdint.h>
 
 #ifndef ARDUINO
 #define TASK_SECOND 1000UL
 #endif
 
 class ITimeLapse {
-    public:
-        virtual ~ITimeLapse() { }
-        virtual unsigned long millis() = 0;
-    };
+public:
+    virtual ~ITimeLapse() { }
+    virtual uint32_t millis() = 0;
+};
 
 class IShellyWrapper : public ITimeLapse {
 public:
