+++ include/LimitControlCalculation.h
@@ -32,16 +32,16 @@ private:
     ShellyClientData& _shellyClientData;
     IShellyWrapper& _shellyWrapper;
 
-    float _invLimitAbsolute{};
-    time_t _intervalPro3em{};
-    time_t _intervalPlugS{};
+    float _invLimitAbsolute {};
+    time_t _intervalPro3em {};
+    time_t _intervalPlugS {};
 
-    float _actLimit{};
-    unsigned long _lastLimitSend{};
+    float _actLimit {};
+    uint32_t _lastLimitSend {};
     std::string _debugPro3em;
     std::string _debugPlugS;
     float _channelPower[INV_MAX_CHAN_COUNT];
-    int _channelCnt{};
+    int _channelCnt {};
 
     CalcLimitFunctionData_t _dataIncrease {};
     CalcLimitFunctionData_t _dataDecrease {};
